Newest version

+++ script/local_scripts/end_to_end.py
@@ -76,18 +76,24 @@ def watch_dir():
         before = dict ([(f, None) for f in os.listdir (path_to_watch)])
         xml_file = minidom.parse("../../xmls_in/"+file)
         itemlist = xml_file.getElementsByTagName('App_Data')
-        s =itemlist[0]
-        content_provider=(s.attributes['Value'].value)
+        if(itemlist):
+            s =itemlist[0]
+            content_provider=(s.attributes['Value'].value)
+        else:
+            content_provider = ''
         if not(content_provider == 'nbcuniversal'):
             content_provider = "viacom"
         else:
             content_provider = "nbcuniversal"
         package_name = file.replace(".xml","")
-        print(content_provider)
         #Make a folder with package name
-        os.makedirs('../../xmls_in/'+package_name)
+        if not os.path.exists('../../xmls_in/'+package_name):
+            os.makedirs('../../xmls_in/'+package_name)
         os.rename("../../xmls_in/" + file, "../../xmls_in/" + package_name +"/" + file) #move file into folder
-        os.rename("../../xmls_in/"+package_name,'../../xmls_out/'+content_provider+'/'+package_name)
+        if not os.path.exists('../../xmls_out/'+content_provider+'/'+package_name):
+            os.rename("../../xmls_in/"+package_name,'../../xmls_out/'+content_provider+'/'+package_name)
+        else:
+            print("Error file already processed")
         return [content_provider,package_name]
     if removed:
         print "Removed: ", ", ".join (removed)
@@ -97,14 +103,13 @@ def pull_from_s3_success(content_provider,package_name):
 # if a file is here that means it was validated, and the package is good to go on to
 # the rest of the gen3 ingest process.
 ### Returns either an array with the file headers and the file content found in the field, or false
-    if(checkLog('gen3-interns-'+content_provider+'total',''+package_name+'.txt')):
+    if(checkLog('gen3-interns-'+content_provider+'total',''+package_name+'_logs.txt')):
         print('New validation info found for success')
         try:
-
-            s3.Bucket('gen3-interns-'+content_provider+'total').download_file(''+package_name+'.txt', '../../xmls_out/'+content_provider+'/'+package_name+'/'+package_name+'_logs.txt') #add LOG to the end
+            s3.Bucket('gen3-interns-'+content_provider+'total').download_file(''+package_name+'_logs.txt', '../../xmls_out/'+content_provider+'/'+package_name+'/'+package_name+'_logs.txt') #add LOG to the end
             os.rename('../../xmls_out/'+content_provider+'/'+package_name , '../../xmls_out/'+content_provider+'/valid/'+package_name) #moves package folder into success or failure
             print('Validation info retrieved from s3, shows a validation success')
-            client.delete_object(Bucket='gen3-interns-'+content_provider+'total', Key ='logs/'+package_name+'.txt')
+            client.delete_object(Bucket='gen3-interns-'+content_provider+'total', Key =''+package_name+'_logs.txt')
             print('Old validation info deleted from s3')
             file = open('../../xmls_out/'+content_provider+'/valid/'+package_name+'/'+package_name+'_logs.txt') #add LOG to the end
             #print(file.read())
@@ -124,10 +129,10 @@ def pull_from_s3_failures(content_provider,package_name):
 # that the xml didn't pass validation and that the content provider and our team must be alerted
 # this function also moves the package to a seperate location while it waits for our team to deal with it
 ### Returns either an array with the file headers and the file content found in the field, or false
-    if(checkLog('gen3-interns-'+content_provider+'failures',''+package_name+'.txt')):
+    if(checkLog('gen3-interns-'+content_provider+'failures',''+package_name+'_logs.txt')):
         print('New validation info found for failure')
         try:
-            s3.Bucket('gen3-interns-'+content_provider+'failures').download_file(''+package_name+'.txt', '../../xmls_out/'+content_provider+'/'+package_name+'/'+package_name+'_logs.txt') #add LOG to the end
+            s3.Bucket('gen3-interns-'+content_provider+'failures').download_file(''+package_name+'_logs.txt', '../../xmls_out/'+content_provider+'/'+package_name+'/'+package_name+'_logs.txt') #add LOG to the end
             os.rename('../../xmls_out/'+content_provider+'/'+package_name , '../../xmls_out/'+content_provider+'/invalid/'+package_name) #moves package folder into success or failure
             print('Validation info retrieved from s3, shows a validation failure')
             print('Validation failed info stored in s3')
@@ -148,7 +153,6 @@ def checkLog(bucket,key):
 # Helper function to check a bucket for the specific filename (key) to see if the valication
 # process is done yet. returns a boolean that is True if the key was found in the bucket
     try:
-        print("b: "+bucket+", k: "+key)
         s3.Object(bucket, key).load()
     except botocore.exceptions.ClientError as e:
         #print('404 file not found')
